Implicit Euler with Newtons Method now working after refactor (not sure about damping)

+++ IntegratorsAbstract.cpp
@@ -97,7 +97,7 @@ void IntegratorAbstract::printInfo(){
 	// /////////////////
 }
 
-void IntegratorAbstract::initializeIntegrator(int ph, SolidMesh& pM, MatrixXd& pTV){
+void IntegratorAbstract::initializeIntegrator(double ph, SolidMesh& pM, MatrixXd& pTV){
 	//Constants
 	vertsNum = pTV.rows();
