+++ Verlet.h
@@ -17,7 +17,7 @@ using namespace std;
 class Verlet: public IntegratorAbstract{
 
 public:
-	void initializeIntegrator(int ph, SolidMesh& pM, MatrixXd& pTV);
+	void initializeIntegrator(double ph, SolidMesh& pM, MatrixXd& pTV);
 	void setXIntoTV(VectorXd& x_new);
 	void createForceVector();
 	void calculateGravity();
